Simplified audio and redid input

+++ gfx/const.go
@@ -47,10 +47,10 @@ var (
 	shaderColor         = gl.Attrib{Value: 2}
 	shaderConstantColor = gl.Attrib{Value: 3}
 
-	shaderPosFlag           = uint32(1 << shaderPos.Value)
-	shaderTexCoordFlag      = uint32(1 << shaderTexCoord.Value)
-	shaderColorFlag         = uint32(1 << shaderColor.Value)
-	shaderConstantColorFlag = uint32(1 << shaderConstantColor.Value)
+	shaderPosFlag           = uint32(1 << uint32(shaderPos.Value))
+	shaderTexCoordFlag      = uint32(1 << uint32(shaderTexCoord.Value))
+	shaderColorFlag         = uint32(1 << uint32(shaderColor.Value))
+	shaderConstantColorFlag = uint32(1 << uint32(shaderConstantColor.Value))
 )
 
 // Draw modes
